Se agrego opcion para listar compras y ventas, ademas se agrego opcion para gasto, y se cambio el dao de gasto

+++ src/main/java/dart/restaurante/controller/CajaJpaController.java
@@ -4,17 +4,20 @@
  */
 package dart.restaurante.controller;
 
+import dart.restaurante.controller.exceptions.IllegalOrphanException;
 import dart.restaurante.controller.exceptions.NonexistentEntityException;
 import dart.restaurante.controller.exceptions.PreexistingEntityException;
 import dart.restaurante.dao.Caja;
 import java.io.Serializable;
-import java.util.List;
-import javax.persistence.EntityManager;
-import javax.persistence.EntityManagerFactory;
 import javax.persistence.Query;
 import javax.persistence.EntityNotFoundException;
 import javax.persistence.criteria.CriteriaQuery;
 import javax.persistence.criteria.Root;
+import dart.restaurante.dao.Gasto;
+import java.util.ArrayList;
+import java.util.List;
+import javax.persistence.EntityManager;
+import javax.persistence.EntityManagerFactory;
 
 /**
  *
@@ -36,7 +39,21 @@ public class CajaJpaController implements Serializable {
         try {
             em = getEntityManager();
             em.getTransaction().begin();
+            Gasto idGasto = caja.getIdGasto();
+            if (idGasto != null) {
+                idGasto = em.getReference(idGasto.getClass(), idGasto.getId());
+                caja.setIdGasto(idGasto);
+            }
             em.persist(caja);
+            if (idGasto != null) {
+                Caja oldIdCajaOfIdGasto = idGasto.getIdCaja();
+                if (oldIdCajaOfIdGasto != null) {
+                    oldIdCajaOfIdGasto.setIdGasto(null);
+                    oldIdCajaOfIdGasto = em.merge(oldIdCajaOfIdGasto);
+                }
+                idGasto.setIdCaja(caja);
+                idGasto = em.merge(idGasto);
+            }
             em.getTransaction().commit();
         } catch (Exception ex) {
             if (findCaja(caja.getId()) != null) {
@@ -50,12 +67,38 @@ public class CajaJpaController implements Serializable {
         }
     }
 
-    public void edit(Caja caja) throws NonexistentEntityException, Exception {
+    public void edit(Caja caja) throws IllegalOrphanException, NonexistentEntityException, Exception {
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
+            Caja persistentCaja = em.find(Caja.class, caja.getId());
+            Gasto idGastoOld = persistentCaja.getIdGasto();
+            Gasto idGastoNew = caja.getIdGasto();
+            List<String> illegalOrphanMessages = null;
+            if (idGastoOld != null && !idGastoOld.equals(idGastoNew)) {
+                if (illegalOrphanMessages == null) {
+                    illegalOrphanMessages = new ArrayList<String>();
+                }
+                illegalOrphanMessages.add("You must retain Gasto " + idGastoOld + " since its idCaja field is not nullable.");
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
+            }
+            if (idGastoNew != null) {
+                idGastoNew = em.getReference(idGastoNew.getClass(), idGastoNew.getId());
+                caja.setIdGasto(idGastoNew);
+            }
             caja = em.merge(caja);
+            if (idGastoNew != null && !idGastoNew.equals(idGastoOld)) {
+                Caja oldIdCajaOfIdGasto = idGastoNew.getIdCaja();
+                if (oldIdCajaOfIdGasto != null) {
+                    oldIdCajaOfIdGasto.setIdGasto(null);
+                    oldIdCajaOfIdGasto = em.merge(oldIdCajaOfIdGasto);
+                }
+                idGastoNew.setIdCaja(caja);
+                idGastoNew = em.merge(idGastoNew);
+            }
             em.getTransaction().commit();
         } catch (Exception ex) {
             String msg = ex.getLocalizedMessage();
@@ -73,7 +116,7 @@ public class CajaJpaController implements Serializable {
         }
     }
 
-    public void destroy(String id) throws NonexistentEntityException {
+    public void destroy(String id) throws IllegalOrphanException, NonexistentEntityException {
         EntityManager em = null;
         try {
             em = getEntityManager();
@@ -84,6 +127,17 @@ public class CajaJpaController implements Serializable {
                 caja.getId();
             } catch (EntityNotFoundException enfe) {
                 throw new NonexistentEntityException("The caja with id " + id + " no longer exists.", enfe);
+            }
+            List<String> illegalOrphanMessages = null;
+            Gasto idGastoOrphanCheck = caja.getIdGasto();
+            if (idGastoOrphanCheck != null) {
+                if (illegalOrphanMessages == null) {
+                    illegalOrphanMessages = new ArrayList<String>();
+                }
+                illegalOrphanMessages.add("This Caja (" + caja + ") cannot be destroyed since the Gasto " + idGastoOrphanCheck + " in its idGasto field has a non-nullable idCaja field.");
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
             }
             em.remove(caja);
             em.getTransaction().commit();
